admin: move admin access token CLI option to `reana_admin`
